requirements.txt
Django==5.2.5
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==11.3.0
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.10
dj-database-url==2.1.0
mysqlclient==2.2.4
